+++ ui/src/components/timestamp.js
@@ -3,18 +3,30 @@ import fromUnixTime from 'date-fns/fromUnixTime';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-export default function Timestamp({ timestamp, past }) {
-  if (!timestamp) return <em>never</em>;
+export default function Timestamp({
+  timestamp,
+  expiration = false,
+  className = undefined,
+}) {
+  if (!timestamp) return <em className={className}>never</em>;
   const date = fromUnixTime(timestamp);
-  const relative = formatDistanceToNow(date, { addSuffix: past });
   const absolute = date.toISOString().replace(/\.0+Z$/, 'Z');
+  if (expiration && date < new Date()) {
+    return (
+      <time title={absolute} dateTime={absolute} className={className}>
+        expired
+      </time>
+    );
+  }
+  const relative = formatDistanceToNow(date, { addSuffix: !expiration });
   return (
-    <time title={absolute} dateTime={absolute}>
+    <time title={absolute} dateTime={absolute} className={className}>
       {relative}
     </time>
   );
 }
 Timestamp.propTypes = {
   timestamp: PropTypes.number.isRequired,
-  past: PropTypes.bool.isRequired,
+  expiration: PropTypes.bool,
+  className: PropTypes.string,
 };