+++ ui/src/components/tokenChangeTable.js
@@ -11,7 +11,7 @@ export default function TokenChangeTable({ data }) {
       {
         Header: 'Event Time',
         // eslint-disable-next-line react/display-name, react/prop-types
-        Cell: ({ value }) => <Timestamp timestamp={value} past />,
+        Cell: ({ value }) => <Timestamp timestamp={value} />,
         accessor: 'event_time',
       },
       {
@@ -63,13 +63,15 @@ export default function TokenChangeTable({ data }) {
       {
         Header: 'Expires',
         // eslint-disable-next-line react/display-name, react/prop-types
-        Cell: ({ value }) => <Timestamp timestamp={value} past={false} />,
+        Cell: ({ value }) => (
+          <Timestamp timestamp={value} className="qa-expires" expiration />
+        ),
         accessor: 'expires',
       },
       {
         Header: 'Old Expires',
         // eslint-disable-next-line react/display-name, react/prop-types
-        Cell: ({ value }) => <Timestamp timestamp={value} past={false} />,
+        Cell: ({ value }) => <Timestamp timestamp={value} expiration />,
         accessor: 'old_expires',
       },
     ],